Split storage handler code out of artifacts.py
